Resolve #8 by increasing bmpDraw parameter size to 16 bits.

+++ examples/spitftbitmap/spitftbitmap.ino
@@ -65,7 +65,7 @@ void loop() {
 
 #define BUFFPIXEL 20
 
-void bmpDraw(char *filename, uint8_t x, uint8_t y) {
+void bmpDraw(char *filename, uint16_t x, uint16_t y) {
 
   File     bmpFile;
   int      bmpWidth, bmpHeight;   // W+H in pixels
